# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title:  DevOps Ecosystem Platform — Combined API
  version: "2.0.0"
  description: |
    API contract for 4 modules added in BRD-001 v1.1 (FR-019 – FR-032).
    Auth: Bearer JWT (ADR-003). All write operations produce audit log (NFR-007).
    Banking-grade: all endpoints except /health require Authorization header.
  contact:
    name: Backend Dev (Agent 06)
servers:
  - url: http://localhost:8001
    description: PPG System — Annual Plans, Project Objects, Reports
  - url: http://localhost:8002
    description: BA Workflow — BA Documents v2
  - url: http://localhost:8003
    description: Test Platform — Test Documents
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/documents/{doc_id}/files:
    post:
      summary: Upload file attachment (ADR-005)
      operationId: uploadBADocFile
      tags: [ba-documents]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded
        "413":
          description: FILE_TOO_LARGE
        "415":
          description: FILE_TYPE_NOT_ALLOWED

    get:
      summary: List current files for document
      operationId: listBADocFiles
      tags: [ba-documents]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of file records
...
  /api/v1/test-documents/{doc_id}/files:
    post:
      summary: Upload file for test document (ADR-005)
      operationId: uploadTestDocFile
      tags: [test-documents]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded
        "413":
          description: FILE_TOO_LARGE
        "415":
          description: FILE_TYPE_NOT_ALLOWED
...